requirements.txt
Django==2.1.1
djangorestframework==3.10.3
numpy==1.17.3
pytz==2019.3
sqlparse==0.3.0
